requirements.txt
fastmcp>=0.1.0
fastapi>=0.110.0
httpx>=0.27.0
anthropic>=0.40.0
python-dotenv>=1.0.0
langgraph>=0.2.0
langchain-anthropic>=0.3.0
pydantic>=2.0.0
uvicorn>=0.29.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
